android/anc/src/main/java/org/smartregister/fhircore/anc/ui/pin/PinSetupActivity.kt
/*
 * Copyright 2021 Ona Systems, Inc
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 *       http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */

package org.smartregister.fhircore.anc.ui.pin

import android.content.Intent
import android.os.Bundle
import android.view.WindowManager
import androidx.activity.compose.setContent
import androidx.activity.viewModels
import dagger.hilt.android.AndroidEntryPoint
import org.smartregister.fhircore.anc.ui.family.register.FamilyRegisterActivity
import org.smartregister.fhircore.engine.ui.appsetting.AppSettingActivity
import org.smartregister.fhircore.engine.ui.base.BaseMultiLanguageActivity
import org.smartregister.fhircore.engine.ui.theme.AppTheme
import org.smartregister.fhircore.engine.util.FORCE_LOGIN_VIA_USERNAME

@AndroidEntryPoint
class PinSetupActivity : BaseMultiLanguageActivity() {

  val pinViewModel by viewModels<PinViewModel>()

  override fun onCreate(savedInstanceState: Bundle?) {
    super.onCreate(savedInstanceState)

    window.setSoftInputMode(WindowManager.LayoutParams.SOFT_INPUT_ADJUST_RESIZE)

    pinViewModel.apply {
      loadData(isSetup = true)
      val pinSetupActivity = this@PinSetupActivity
      navigateToHome.observe(pinSetupActivity) { pinSetupActivity.moveToHome() }
      navigateToSettings.observe(pinSetupActivity) { pinSetupActivity.moveToSettings() }
      pin.observe(pinSetupActivity) { it.let { enableSetPin.postValue(it.length > 3) } }
    }
    setContent { AppTheme { PinSetupScreen(pinViewModel) } }
  }

  private fun moveToHome() {
    sharedPreferencesHelper.write(FORCE_LOGIN_VIA_USERNAME, "false")
    startActivity(
      Intent(this, FamilyRegisterActivity::class.java).apply {
        addFlags(Intent.FLAG_ACTIVITY_NEW_TASK)
      }
    )
    finish()
  }

  private fun moveToSettings() {
    startActivity(
      Intent(this, AppSettingActivity::class.java).apply {
        addFlags(Intent.FLAG_ACTIVITY_CLEAR_TOP)
        addFlags(Intent.FLAG_ACTIVITY_NO_HISTORY)
        addFlags(Intent.FLAG_ACTIVITY_SINGLE_TOP)
        addFlags(Intent.FLAG_ACTIVITY_NEW_TASK)
        addCategory(Intent.CATEGORY_LAUNCHER)
      }
    )
    finish()
  }
}
